refactor: add List component for occasionList (#159)
