--- src/fsd_entities/ocaasion/config/occasionRequestManagementTypes.ts
@@ -1,15 +1,0 @@
-interface OccasionRequestManagementProps {
-  state: string | undefined;
-  title: string;
-  firstNavigation: {
-    name: string;
-    state: string;
-    router: string;
-  };
-  navigation?: {
-    name: string;
-    state: string;
-    router: string;
-  }[];
-  isFirstNavigation: boolean;
-}
