+++ src/fsd_widgets/occasion/ui/index.ts
@@ -1,1 +1,2 @@
 export { OccasionRequestManagement } from './OccasionRequestManagement';
+export { OccasionList } from './OccasionList';
